requirements.txt
lmdb==1.4.1
tqdm==4.66.1
datasets
tensorboardX
accelerate
ninja
transformers==4.46.3
pycocotools==2.0.7
scikit-image
Pillow==9.5.0
opencv-python-headless
datasets
einops==0.8.0
sentencepiece
xformers==0.0.29
pydantic==2.9.2
deepspeed
peft==0.14.0
diffusers==0.32.0
rotary-embedding-torch==0.8.4
tiktoken==0.8.0
transformers_stream_generator==0.0.5
ftfy
bs4
bson==0.5.10
gradio==3.50.2
httpx==0.23.3
fairscale==0.4.13
kornia
timm==1.0.9
protobuf==3.20.0
basicsr
sentencepiece
huggingface_hub
prodigyopt